wlutz-arc working again
