+++ lib/pymedphys/_experimental/streamlit/apps/iview_to_dicom.py
@@ -12,12 +12,13 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-
 from pymedphys._imports import streamlit as st
 
+import pymedphys._dicom.create as _pp_dcm_create
+from pymedphys import _losslessjpeg as lljpeg
 from pymedphys._streamlit import categories
 
-from .wlutz import _config, _utilities
+# from .wlutz import _config, _utilities
 
 CATEGORY = categories.PLANNING
 TITLE = "iView to DICOM"
@@ -35,3 +36,53 @@ def main():
         selected_date,
         selected_machine_id,
     ) = _utilities.get_directories_and_initial_database(config, refresh_cache)
+
+
+def _create_portal_image_dicom_dataset(
+    gantry_angle, collimator_angle, table_angle, image_path
+):
+    """Don't intend this DICOM file to be compliant to the spec.
+
+    Instead, for now, just enough that software such as PIPs will
+    happily accept the created file.
+    """
+
+    # TODO: Refactor this so that for all pylinac calls a full DICOM
+    # file is passed to it in the way that it expects. The image wrapper
+    # around pylinac could instead call this first.
+
+    # Image plane pixel spacing.
+    # Exported DICOM files have an image plane pixel spacing of 0.405,
+    # with SID of 1600 and SAD of 1000. This corresponds to an iso
+    # centre pixel spacing of 0.2531 mm.
+    #
+    # Within the database, those same images have an isocentre pixel
+    # spacing of either 0.2510 mm or 0.2488 mm. I suspect potentially
+    # there might be some interesting datastore rounding going on here.
+    # I was under the impression that the isocentre pixel spacing was
+    # for these images was 0.25 mm.
+
+    pixel_array = lljpeg.imread(image_path)
+    pixels_per_mm = _pp_wlutz_iview.infer_pixels_per_mm_from_shape(pixel_array)
+
+    sid = 1600.0
+    sad = 1000.0
+
+    pixel_spacing = 1 / pixels_per_mm * sid / sad
+
+    ds = _pp_dcm_create.dicom_dataset_from_dict(
+        {
+            "ImageType": ["ORIGINAL", "PRIMARY", "PORTAL"],
+            "Rows": pixel_array.shape[0],
+            "Columns": pixel_array.shape[1],
+            "GantryAngle": gantry_angle,
+            "BeamLimitingDeviceAngle ": collimator_angle,
+            "PatientSupportAngle ": table_angle,
+            "PixelData": pixel_array,
+            "RadiationMachineSAD": sad,
+            "RTImageSID": sid,
+            "ImagePlanePixelSpacing": [pixel_spacing, pixel_spacing],
+        }
+    )
+
+    return ds
